code/requirements.txt
requests
elasticsearch
jinja2
PyGithub
pytz
linkedin-api-client
minio
feedparser